OMJulia.API

Module OMJulia.API aims to provide a Julia interface to the OpenModelica scripting API. In contrast to sending the scripting api calls directly via sendExpression this API has a Julia-like interface and some level of error handling implemented. This means errors will throw Julia Exception OMJulia.API.ScriptingError instead of only printing to stdout.

Warn

Not all OMJulia.API functions are tested and some functions could have slightly different default values compared to the OpenModelica scripting API.

Instead of escaping strings yourself the API interface handles this for you:

sendExpression(omc, "loadFile(\"$(bouncingBallFile)\")")

becomes

API.loadFile(omc, bouncingBallFile)

Functions

OMJulia.API.buildModel — Method

buildModel(omc, className;
           startTime = 0.0,
           stopTime = 1.0,
           numberOfIntervals = 500,
           tolerance = 1e-6,
           method = "",
           fileNamePrefix = className,
           options = "",
           outputFormat = "mat",
           variableFilter = ".*",
           cflags = "",
           simflags = "")

Build Modelica model by generating C code and compiling it into an executable simulation. It does not run the simulation!

See OpenModelica scripting API buildModel.

OMJulia.API.buildModelFMU — Method

buildModelFMU(omc, className;
              version = "2.0",
              fmuType = "me",
              fileNamePrefix=className,
              platforms=["static"],
              includeResources = false)

Translates a modelica model into a Functional Mockup Unit. The only required argument is the className, while all others have some default values.

See OpenModelica scripting API buildModelFMU.

OMJulia.API.cd — Function

cd(omc, newWorkingDirectory="")

Change directory to the given path newWorkingDirectory (which may be either relative or absolute). Returns the new working directory on success or a message on failure. If the given path is the empty string, the function simply returns the current working directory.

See OpenModelica scripting API cd.

OMJulia.API.closeSimulationResultFile — Method

    closeSimulationResultFile(omc)

Closes the current simulation result file. Only needed by Windows. Windows cannot handle reading and writing to the same file from different processes. To allow OMEdit to make successful simulation again on the same file we must close the file after reading the Simulation Result Variables. Even OMEdit only use this API for Windows.

See OpenModelica scripting API closeSimulationResultFile.

OMJulia.API.diffSimulationResults — Method

diffSimulationResults(omc, actualFile, expectedFile, diffPrefix;
                      relTol = 1e-3,
                      relTolDiffMinMax = 1e-4,
                      rangeDelta = 0.002,
                      vars = String[],
                      keepEqualResults = false)

Compares simulation results.

See OpenModelica scripting API diffSimulationResults.

OMJulia.API.getClassNames — Method

getClassNames(omc;
              class_ = "",
              recursive = false,
              qualified = false,
              sort = false,
              builtin = false,
              showProtected = false,
              includeConstants = false)

Returns the list of class names defined in the class.

See OpenModelica scripting API getClassNames.

OMJulia.API.getErrorString — Method

getErrorString(omc, warningsAsErrors = false)

Returns the current error message.

See OpenModelica scripting API getErrorString.

OMJulia.API.getInstallationDirectoryPath — Method

getInstallationDirectoryPath(omc)

This returns OPENMODELICAHOME if it is set; on some platforms the default path is returned if it is not set.

See OpenModelica scripting API getInstallationDirectoryPath.

OMJulia.API.getVersion — Method

getVersion(omc)

Returns the version of the Modelica compiler.

See OpenModelica scripting API getVersion.

OMJulia.API.linearize — Method

Creates a model with symbolic linearization matrices.

See OpenModelica scripting API linearize.

OMJulia.API.loadFile — Method

loadFile(omc, fileName;
         encoding = "",
         uses = true,
         notify = true,
         requireExactVersion = false)

Load file fileName (*.mo) and merge it with the loaded AST. See OpenModelica scripting API loadFile.

OMJulia.API.loadModel — Method

loadModel(omc, className;
          priorityVersion = String[],
          notify = false,
          languageStandard = "",
          requireExactVersion = false)

Loads a Modelica library.

See OpenModelica scripting API loadModel.

OMJulia.API.makeVectorString — Method

makeVectorString(vec)

Add quotes around each string element.

OMJulia.API.readSimulationResult — Function

readSimulationResult(omc, filename,
                     variables = String[],
                     size = 0)

Reads a result file, returning a matrix corresponding to the variables and size given.

See OpenModelica scripting API readSimulationResult.

OMJulia.API.readSimulationResultSize — Method

readSimulationResultSize(omc, fileName)

The number of intervals that are present in the output file.

See OpenModelica scripting API readSimulationResultSize.

OMJulia.API.readSimulationResultVars — Method

readSimulationResultVars(omc, fileName; readParameters = true, openmodelicaStyle = false)

Returns the variables in the simulation file; you can use val() and plot() commands using these names.

See OpenModelica scripting API readSimulationResultVars.

OMJulia.API.setCommandLineOptions — Method

setCommandLineOptions(omc, option)

The input is a regular command-line flag given to OMC, e.g. -d=failtrace or -g=MetaModelica.

See OpenModelica scripting API setCommandLineOptions.

OMJulia.API.simulate — Method

simulate(omc, className;
        startTime = 0.0,
        stopTime = nothing,
        numberOfIntervals = 500,
        tolerance = 1e-6,
        method = "",
        fileNamePrefix=className,
        options = "",
        outputFormat = "mat",
        variableFilter = ".*",
        cflags = "",
        simflags = "")

Simulates a modelica model by generating C code, build it and run the simulation executable.

See OpenModelica scripting API simulate.

OMJulia.API.ScriptingError — Type

ScriptingError <: Exception

OpenModelica scripting error with message msg and additional error string from getErrroString.