Advanced API

ModelicaSystem

OMJulia.ModelicaSystem — Function

ModelicaSystem(omc, filename, modelname, library=nothing;
               actcommandLineOptions=nothing, variableFilter=nothing)

Set command line options for OMCSession and build model modelname to prepare for a simulation.

Arguments

omc: OpenModelica compiler session, see OMCSession().

filename: Path to Modelica file.

modelname: Name of Modelica model to build, including namespace if the model is wrappen within a Modelica package.

library: List of dependent libraries or Modelica files. This argument can be passed as string (e.g. "Modelica") or tuple (e.g. ("Modelica", "4.0") or array (e.g. ["Modelica", "SystemDynamics"] or [("Modelica", "4.0"), "SystemDynamics"]).

Keyword Arguments

commandLineOptions: OpenModelica command line options, see OpenModelica Compiler Flags.

variableFilter: Regex to filter variables in result file.

Usage

using OMJulia
mod = OMJulia.OMCSession()
ModelicaSystem(mod, "BouncingBall.mo", "BouncingBall", ["Modelica", "SystemDynamics"], commandLineOptions="-d=newInst")

Providing dependent libaries:

using OMJulia
mod = OMJulia.OMCSession()
ModelicaSystem(mod, "BouncingBall.mo", "BouncingBall", ["Modelica", "SystemDynamics", "dcmotor.mo"])

See also OMCSession().

OMJulia.OMCSession — Type

OMCSession(omc=nothing)

Create new OpenModelica session.

Arguments

omc: "Path to OpenModelica compiler"

See also ModelicaSystem.

Example

julia> using OMJulia
julia> mod = OMJulia.OMCSession()[ Info: Path to zmq file="/tmp/openmodelica.runner.port.julia.HpykbEHzg6"
OMJulia.OMCSession(false, false, Dict{Any, Any}(), Dict{Any, Any}(), Dict{Any, Any}(), "", "/home/runner/work/OMJulia.jl/OMJulia.jl", "", "", "", "", "", "", Any[], Dict{Any, Any}(), Dict{Any, Any}(), Dict{Any, Any}(), Dict{Any, Any}(), Dict("startTime" => "0.0", "stopTime" => "1.0", "stepSize" => "0.002", "tolerance" => "1e-6"), "", false, "", #undef, #undef, #undef, #undef, ZMQ.Context(Ptr{Nothing} @0x0000000001ffdb70, WeakRef[WeakRef(ZMQ.Socket(Ptr{Nothing} @0x0000000001fba410, FileWatching._FDWatcher(Ptr{Nothing} @0x00000000024a6310, 30, (1, 0), Base.GenericCondition{Base.Threads.SpinLock}(Base.IntrusiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0)), 0, (false, false))))]), ZMQ.Socket(Ptr{Nothing} @0x0000000001fba410, FileWatching._FDWatcher(Ptr{Nothing} @0x00000000024a6310, 30, (1, 0), Base.GenericCondition{Base.Threads.SpinLock}(Base.IntrusiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0)), 0, (false, false))), Process(`omc --interactive=zmq -z=julia.HpykbEHzg6`, ProcessRunning))
julia> ModelicaSystem(mod,
                      joinpath("docs", "testmodels", "BouncingBall.mo"),
                      "BouncingBall")

WorkDirectory

For each OMJulia session a temporary work directory is created and the results are published in that working directory. In order to get the work directory use getWorkDirectory.

OMJulia.getWorkDirectory — Function

Function which returns the working directory of current OMJulia Session for each session a temporary directory is created and the simulation results are generated

Build Model

In case the Modelica model needs to be updated or additional simulation flags needs to be set using sendExpression The buildModel API can be used after ModelicaSystem.

OMJulia.buildModel — Function

Standard buildModel API which builds the modelica model

Get Methods

OMJulia.getQuantities — Function

standard getXXX() API function which return list of all variables parsed from xml file

OMJulia.showQuantities — Function

standard getXXX() API function same as getQuantities(), but returns all the variables as table

OMJulia.getContinuous — Function

standard getXXX() API function which returns the continuous variables parsed from xmlfile

OMJulia.getInputs — Function

standard getXXX() API function which returns the input variables parsed from xmlfile

OMJulia.getOutputs — Function

standard getXXX() API function which returns the output variables parsed from xmlfile

OMJulia.getParameters — Function

standard getXXX() API function which returns the parameter variables parsed from xmlfile

OMJulia.getSimulationOptions — Function

standard getXXX() API function which returns the SimulationOption variables parsed from xmlfile

OMJulia.getSolutions — Function

standard getXXX() API Function which reads the result file and return the simulation results to user which can be used for plotting or further anlaysis

Examples

julia> getQuantities(mod)13-element Vector{Any}:
 Dict{Any, Any}("name" => "h", "max" => "None", "start" => "1.0", "min" => "None", "aliasvariable" => nothing, "variability" => "continuous", "changeable" => "true", "causality" => "local", "description" => "height of ball", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "v", "max" => "None", "start" => "None", "min" => "None", "aliasvariable" => nothing, "variability" => "continuous", "changeable" => "true", "causality" => "local", "description" => "velocity of ball", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "der(h)", "max" => "None", "start" => "None", "min" => "None", "aliasvariable" => nothing, "variability" => "continuous", "changeable" => "false", "causality" => "local", "description" => "der(height of ball)", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "der(v)", "max" => "None", "start" => "None", "min" => "None", "aliasvariable" => nothing, "variability" => "continuous", "changeable" => "false", "causality" => "local", "description" => "der(velocity of ball)", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "v_new", "max" => "None", "start" => "None", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => nothing, "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "e", "max" => "None", "start" => "0.7", "min" => "None", "aliasvariable" => nothing, "variability" => "parameter", "changeable" => "true", "causality" => "parameter", "description" => "coefficient of restitution", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "g", "max" => "None", "start" => "9.81", "min" => "None", "aliasvariable" => nothing, "variability" => "parameter", "changeable" => "true", "causality" => "parameter", "description" => "gravity acceleration", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "foo", "max" => "None", "start" => "2", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => nothing, "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "\$whenCondition1", "max" => "None", "start" => "None", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => "impact and v <= 0.0", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "\$whenCondition2", "max" => "None", "start" => "false", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => "impact", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "\$whenCondition3", "max" => "None", "start" => "None", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => "h <= 0.0 and v <= 0.0", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "flying", "max" => "None", "start" => "true", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => "true, if ball is flying", "alias" => "noAlias"…)
 Dict{Any, Any}("name" => "impact", "max" => "None", "start" => "false", "min" => "None", "aliasvariable" => nothing, "variability" => "discrete", "changeable" => "false", "causality" => "local", "description" => nothing, "alias" => "noAlias"…)
julia> getQuantities(mod, "height")Any[]
julia> getQuantities(mod, ["c","radius"])Any[]
julia> showQuantities(mod)13×10 DataFrame
 Row │ max     description                 name             start   variabilit ⋯
     │ String  Union…                      String           String  String     ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │ None    height of ball              h                1.0     continuous ⋯
   2 │ None    velocity of ball            v                None    continuous
   3 │ None    der(height of ball)         der(h)           None    continuous
   4 │ None    der(velocity of ball)       der(v)           None    continuous
   5 │ None                                v_new            None    discrete   ⋯
   6 │ None    coefficient of restitution  e                0.7     parameter
   7 │ None    gravity acceleration        g                9.81    parameter
   8 │ None                                foo              2       discrete
   9 │ None    impact and v <= 0.0         $whenCondition1  None    discrete   ⋯
  10 │ None    impact                      $whenCondition2  false   discrete
  11 │ None    h <= 0.0 and v <= 0.0       $whenCondition3  None    discrete
  12 │ None    true, if ball is flying     flying           true    discrete
  13 │ None                                impact           false   discrete   ⋯
                                                               6 columns omitted

julia> getContinuous(mod)Dict{Any, Any} with 4 entries:
  "v"      => "None"
  "h"      => "1.0"
  "der(v)" => "None"
  "der(h)" => "None"
julia> getContinuous(mod, ["velocity","height"])2-element Vector{Int64}:
 0
 0

julia> getInputs(mod)Dict{Any, Any}()
julia> getOutputs(mod)Dict{Any, Any}()

julia> getParameters(mod)Dict{Any, Any} with 2 entries:
  "g" => "9.81"
  "e" => "0.7"
julia> getParameters(mod, ["c","radius"])2-element Vector{Int64}:
 0
 0

julia> getSimulationOptions(mod)Dict{Any, Any} with 5 entries:
  "startTime" => "0"
  "stopTime"  => "1"
  "solver"    => "dassl"
  "stepSize"  => "0.002"
  "tolerance" => "1e-06"
julia> getSimulationOptions(mod, ["stepSize","tolerance"])2-element Vector{String}:
 "0.002"
 "1e-06"

julia> getSolutions(mod)ResultFile does not exist !/home/runner/work/OMJulia.jl/OMJulia.jl/
julia> getSolutions(mod, ["time","height"])ResultFile does not exist !/home/runner/work/OMJulia.jl/OMJulia.jl/

Set Methods

Missing docstring.

Missing docstring for setInputs(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for setParameters(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for setSimulationOptions(). Check Documenter's build log for details.

Examples

julia> setInputs(mod, "cAi=1")cAi  is not a Input
julia> setInputs(mod, ["cAi=1","Ti=2"])cAi  is not a Input

julia> setParameters(mod, "radius=14")| info |  setParameters() failed: "radius" is not a parameter
julia> setParameters(mod, ["radius=14","c=0.5"])| info |  setParameters() failed: "radius" is not a parameter
| info |  setParameters() failed: "c" is not a parameter

julia> setSimulationOptions(mod, ["stopTime=2.0","tolerance=1e-08"])

Simulation

OMJulia.simulate — Function

simulate(omc; resultfile=nothing, simflags=nothing, verbose=true)

Simulate modelica model.

Arguments

omc: OpenModelica compiler session, see OMCSession().

Keyword Arguments

resultFile: Result file to write simulation results into.

simflags: Simulation flags, see Simulation Runtime Flags.

Examples

simulate(omc)

Specify result file:

simulate(omc, resultfile="tmpresult.mat")

Set simulation runtime flags:

simulate(omc, simflags="-noEmitEvent -override=e=0.3,g=9.3")

Examples

julia> simulate(mod, simflags="-noEventEmit -noRestart -override=e=0.3,g=9.71")stopTime=2.0

tolerance=1e-08

LOG_STDOUT        | warning | invalid command line option: -noEventEmit -noRestart -override=e=0.3,g=9.71
LOG_STDOUT        | info    | usage: /tmp/jl_Zs01Eq/BouncingBall
|                 | |       | | <-abortSlowSimulation>
|                 | |       | |   aborts if the simulation chatters
|                 | |       | | <-alarm=value> or <-alarm value>
|                 | |       | |   aborts after the given number of seconds (0 disables)
|                 | |       | | <-clock=value> or <-clock value>
|                 | |       | |   selects the type of clock to use -clock=RT, -clock=CYC or -clock=CPU
|                 | |       | | <-cpu>
|                 | |       | |   dumps the cpu-time into the result file
|                 | |       | | <-csvOstep=value> or <-csvOstep value>
|                 | |       | |   value specifies csv-files for debug values for optimizer step
|                 | |       | | <-cvodeNonlinearSolverIteration=value> or <-cvodeNonlinearSolverIteration value>
|                 | |       | |   nonlinear solver iteration for CVODE solver
|                 | |       | | <-cvodeLinearMultistepMethod=value> or <-cvodeLinearMultistepMethod value>
|                 | |       | |   linear multistep method for CVODE solver
|                 | |       | | <-cx=value> or <-cx value>
|                 | |       | |   value specifies a csv-file with inputs as correlation coefficient matrix Cx for DataReconciliation
|                 | |       | | <-daeMode>
|                 | |       | |   flag to let the integrator use daeResiduals
|                 | |       | | <-deltaXLinearize=value> or <-deltaXLinearize value>
|                 | |       | |   value specifies the delta x value for numerical differentiation used by linearization. The default value is 1e-5.
|                 | |       | | <-deltaXSolver=value> or <-deltaXSolver value>
|                 | |       | |   value specifies the delta x value for numerical differentiation used by integrator. The default values is sqrt(DBL_EPSILON).
|                 | |       | | <-embeddedServer=value> or <-embeddedServer value>
|                 | |       | |   enables an embedded server. Valid values: none, opc-da [broken], opc-ua [experimental], or the path to a shared object.
|                 | |       | | <-embeddedServerPort=value> or <-embeddedServerPort value>
|                 | |       | |   [int (default 4841)] value specifies the port number used by the embedded server
|                 | |       | | <-mat_sync=value> or <-mat_sync value>
|                 | |       | |   [int (default 0)] syncs the mat file header after emitting every N time-points (default disabled)
|                 | |       | | <-emit_protected>
|                 | |       | |   emits protected variables to the result-file
|                 | |       | | <-eps=value> or <-eps value>
|                 | |       | |   value specifies the number of convergence iteration to be performed for DataReconciliation
|                 | |       | | <-f=value> or <-f value>
|                 | |       | |   value specifies a new setup XML file to the generated simulation code
|                 | |       | | <-help=value> or <-help value>
|                 | |       | |   get detailed information that specifies the command-line flag
|                 | |       | | <-homAdaptBend=value> or <-homAdaptBend value>
|                 | |       | |   [double (default 0.5)] maximum trajectory bending to accept the homotopy step
|                 | |       | | <-homBacktraceStrategy=value> or <-homBacktraceStrategy value>
|                 | |       | |   value specifies the backtrace strategy in the homotopy corrector step (fix (default), orthogonal)
|                 | |       | | <-homHEps=value> or <-homHEps value>
|                 | |       | |   [double (default 1e-5)] tolerance respecting residuals for the homotopy H-function
|                 | |       | | <-homMaxLambdaSteps=value> or <-homMaxLambdaSteps value>
|                 | |       | |   [int (default size dependent)] maximum lambda steps allowed to run the homotopy path
|                 | |       | | <-homMaxNewtonSteps=value> or <-homMaxNewtonSteps value>
|                 | |       | |   [int (default 20)] maximum newton steps in the homotopy corrector step
|                 | |       | | <-homMaxTries=value> or <-homMaxTries value>
|                 | |       | |   [int (default 10)] maximum number of tries for one homotopy lambda step
|                 | |       | | <-homNegStartDir>
|                 | |       | |   start to run along the homotopy path in the negative direction
|                 | |       | | <-homotopyOnFirstTry>
|                 | |       | |   directly use the homotopy method to solve the initialization problem
|                 | |       | | <-noHomotopyOnFirstTry>
|                 | |       | |   disable the use of the homotopy method to solve the initialization problem
|                 | |       | | <-homTauDecFac=value> or <-homTauDecFac value>
|                 | |       | |   [double (default 10.0)] decrease homotopy step size tau by this factor if tau is too big in the homotopy corrector step
|                 | |       | | <-homTauDecFacPredictor=value> or <-homTauDecFacPredictor value>
|                 | |       | |   [double (default 2.0)] decrease homotopy step size tau by this factor if tau is too big in the homotopy predictor step
|                 | |       | | <-homTauIncFac=value> or <-homTauIncFac value>
|                 | |       | |   [double (default 2.0)] increase homotopy step size tau by this factor if tau is too small in the homotopy corrector step
|                 | |       | | <-homTauIncThreshold=value> or <-homTauIncThreshold value>
|                 | |       | |   [double (default 10.0)] increase the homotopy step size tau if bend < homAdaptBend/homTauIncThreshold
|                 | |       | | <-homTauMax=value> or <-homTauMax value>
|                 | |       | |   [double (default 10.0)] maximum homotopy step size tau for the homotopy process
|                 | |       | | <-homTauMin=value> or <-homTauMin value>
|                 | |       | |   [double (default 1e-4)] minimum homotopy step size tau for the homotopy process
|                 | |       | | <-homTauStart=value> or <-homTauStart value>
|                 | |       | |   [double (default 0.2)] homotopy step size tau at the beginning of the homotopy process
|                 | |       | | <-idaMaxErrorTestFails=value> or <-idaMaxErrorTestFails value>
|                 | |       | |   value specifies the maximum number of error test failures in attempting one step. The default value is 7.
|                 | |       | | <-idaMaxNonLinIters=value> or <-idaMaxNonLinIters value>
|                 | |       | |   value specifies the maximum number of nonlinear solver iterations at one step. The default value is 3.
|                 | |       | | <-idaMaxConvFails=value> or <-idaMaxConvFails value>
|                 | |       | |   value specifies the maximum number of nonlinear solver convergence failures at one step. The default value is 10.
|                 | |       | | <-idaNonLinConvCoef=value> or <-idaNonLinConvCoef value>
|                 | |       | |   value specifies the safety factor in the nonlinear convergence test. The default value is 0.33.
|                 | |       | | <-idaLS=value> or <-idaLS value>
|                 | |       | |   select the linear solver used by ida
|                 | |       | | <-idaScaling>
|                 | |       | |   enable scaling of the IDA solver
|                 | |       | | <-idaSensitivity>
|                 | |       | |   flag to add sensitivity information to the result files
|                 | |       | | <-ignoreHideResult>
|                 | |       | |   ignore HideResult=true annotation
|                 | |       | | <-iif=value> or <-iif value>
|                 | |       | |   value specifies an external file for the initialization of the model relative to -inputPath
|                 | |       | | <-iim=value> or <-iim value>
|                 | |       | |   value specifies the initialization method
|                 | |       | | <-iit=value> or <-iit value>
|                 | |       | |   [double] value specifies a time for the initialization of the model
|                 | |       | | <-ils=value> or <-ils value>
|                 | |       | |   [int (default 3)] number of lambda steps for homotopy methods
|                 | |       | | <-impRKOrder=value> or <-impRKOrder value>
|                 | |       | |   [int (default 5)] value specifies the integration order of the implicit Runge-Kutta method. Valid values: 1-6
|                 | |       | | <-impRKLS=value> or <-impRKLS value>
|                 | |       | |   selects the linear solver of the integration methods: impeuler, trapezoid and imprungekuta
|                 | |       | | <-initialStepSize=value> or <-initialStepSize value>
|                 | |       | |   value specifies an initial step size for supported solver
|                 | |       | | <-csvInput=value> or <-csvInput value>
|                 | |       | |   value specifies an csv-file with inputs for the simulation/optimization of the model
|                 | |       | | <-stateFile=value> or <-stateFile value>
|                 | |       | |   value specifies an file with states start values for the optimization of the model
|                 | |       | | <-inputPath=value> or <-inputPath value>
|                 | |       | |   value specifies a path for reading the input files i.e., model_init.xml and model_info.json
|                 | |       | | <-ipopt_hesse=value> or <-ipopt_hesse value>
|                 | |       | |   value specifies the hessian for Ipopt
|                 | |       | | <-ipopt_init=value> or <-ipopt_init value>
|                 | |       | |   value specifies the initial guess for optimization
|                 | |       | | <-ipopt_jac=value> or <-ipopt_jac value>
|                 | |       | |   value specifies the Jacobian for Ipopt
|                 | |       | | <-ipopt_max_iter=value> or <-ipopt_max_iter value>
|                 | |       | |   value specifies the max number of iteration for ipopt
|                 | |       | | <-ipopt_warm_start=value> or <-ipopt_warm_start value>
|                 | |       | |   value specifies lvl for a warm start in ipopt: 1,2,3,...
|                 | |       | | <-jacobian=value> or <-jacobian value>
|                 | |       | |   select the calculation method of the Jacobian used only by ida and dassl solver.
|                 | |       | | <-jacobianThreads=value> or <-jacobianThreads value>
|                 | |       | |   [int default: 1] value specifies the number of threads for jacobian evaluation in dassl or ida.
|                 | |       | | <-l=value> or <-l value>
|                 | |       | |   value specifies a time where the linearization of the model should be performed
|                 | |       | | <-l_datarec>
|                 | |       | |   emit data recovery matrices with model linearization
|                 | |       | | <-logFormat=value> or <-logFormat value>
|                 | |       | |   value specifies the log format of the executable. -logFormat=text (default), -logFormat=xml or -logFormat=xmltcp
|                 | |       | | <-ls=value> or <-ls value>
|                 | |       | |   value specifies the linear solver method (default: lapack, totalpivot (fallback))
|                 | |       | | <-ls_ipopt=value> or <-ls_ipopt value>
|                 | |       | |   value specifies the linear solver method for ipopt
|                 | |       | | <-lss=value> or <-lss value>
|                 | |       | |   value specifies the linear sparse solver method (default: umfpack)
|                 | |       | | <-lssMaxDensity=value> or <-lssMaxDensity value>
|                 | |       | |   [double (default 0.2)] value specifies the maximum density for using a linear sparse solver
|                 | |       | | <-lssMinSize=value> or <-lssMinSize value>
|                 | |       | |   [int (default 1000)] value specifies the minimum system size for using a linear sparse solver
|                 | |       | | <-lv=value> or <-lv value>
|                 | |       | |   [string list] value specifies the logging level
|                 | |       | | <-lvMaxWarn=value> or <-lvMaxWarn value>
|                 | |       | |   [int (default 3)] maximum times repeating warnings will be displayed
|                 | |       | | <-lv_time=value> or <-lv_time value>
|                 | |       | |   [double list] specifying time interval to allow loging in
|                 | |       | | <-mbi=value> or <-mbi value>
|                 | |       | |   [int (default 0)] value specifies the maximum number of bisection iterations for state event detection or zero for default behavior
|                 | |       | | <-mei=value> or <-mei value>
|                 | |       | |   [int (default 20)] value specifies the maximum number of event iterations
|                 | |       | | <-maxIntegrationOrder=value> or <-maxIntegrationOrder value>
|                 | |       | |   value specifies maximum integration order for supported solver
|                 | |       | | <-maxStepSize=value> or <-maxStepSize value>
|                 | |       | |   value specifies maximum absolute step size for supported solver
|                 | |       | | <-measureTimePlotFormat=value> or <-measureTimePlotFormat value>
|                 | |       | |   value specifies the output format of the measure time functionality
|                 | |       | | <-newtonDiagnostics>
|                 | |       | |   Run newton diagnostics to improve start values for non-linear systems
|                 | |       | | <-newtonFTol=value> or <-newtonFTol value>
|                 | |       | |   [double (default 1e-12)] tolerance respecting residuals for updating solution vector in Newton solver
|                 | |       | | <-newtonMaxStepFactor=value> or <-newtonMaxStepFactor value>
|                 | |       | |   [double (default 1e12)] maximum newton step factor mxnewtstep = maxStepFactor * norm2(xScaling). Used currently only by KINSOL.
|                 | |       | | <-newtonXTol=value> or <-newtonXTol value>
|                 | |       | |   [double (default 1e-12)] tolerance respecting newton correction (delta_x) for updating solution vector in Newton solver
|                 | |       | | <-newton=value> or <-newton value>
|                 | |       | |   value specifies the damping strategy for the newton solver
|                 | |       | | <-nls=value> or <-nls value>
|                 | |       | |   value specifies the nonlinear solver
|                 | |       | | <-nlsInfo>
|                 | |       | |   outputs detailed information about solving process of non-linear systems into csv files.
|                 | |       | | <-nlsLS=value> or <-nlsLS value>
|                 | |       | |   value specifies the linear solver used by the non-linear solver
|                 | |       | | <-nlssMaxDensity=value> or <-nlssMaxDensity value>
|                 | |       | |   [double (default 0.1)] value specifies the maximum density for using a non-linear sparse solver
|                 | |       | | <-nlssMinSize=value> or <-nlssMinSize value>
|                 | |       | |   [int (default 1000)] value specifies the minimum system size for using a non-linear sparse solver
|                 | |       | | <-noemit>
|                 | |       | |   do not emit any results to the result file
|                 | |       | | <-noEquidistantTimeGrid>
|                 | |       | |   stores results not in equidistant time grid as given by stepSize or numberOfIntervals, instead the variable step size of dassl or ida integrator.
|                 | |       | | <-noEquidistantOutputFrequency=value> or <-noEquidistantOutputFrequency value>
|                 | |       | |   value controls the output frequency in noEquidistantTimeGrid mode
|                 | |       | | <-noEquidistantOutputTime=value> or <-noEquidistantOutputTime value>
|                 | |       | |   value controls the output time point in noEquidistantOutputTime mode
|                 | |       | | <-noEventEmit>
|                 | |       | |   do not emit event points to the result file
|                 | |       | | <-noRestart>
|                 | |       | |   disables the restart of the integration method after an event is performed, used by the methods: dassl, ida
|                 | |       | | <-noRootFinding>
|                 | |       | |   disables the internal root finding procedure of methods: dassl and ida.
|                 | |       | | <-noScaling>
|                 | |       | |   disables scaling for the variables and the residuals in the algebraic nonlinear solver KINSOL.
|                 | |       | | <-noSuppressAlg>
|                 | |       | |   flag to not suppress algebraic variables in the local error test of ida solver in daeMode
|                 | |       | | <-optDebugJac=value> or <-optDebugJac value>
|                 | |       | |   value specifies the number of iter from the dyn. optimization, which will be debug, creating *csv and *py file
|                 | |       | | <-optimizerNP=value> or <-optimizerNP value>
|                 | |       | |   value specifies the number of points in a subinterval
|                 | |       | | <-optimizerTimeGrid=value> or <-optimizerTimeGrid value>
|                 | |       | |   value specifies external file with time points.
|                 | |       | | <-output=value> or <-output value>
|                 | |       | |   output the variables a, b and c at the end of the simulation to the standard output
|                 | |       | | <-outputPath=value> or <-outputPath value>
|                 | |       | |   value specifies a path for writing the output files i.e., model_res.mat, model_prof.intdata, model_prof.realdata etc.
|                 | |       | | <-override=value> or <-override value>
|                 | |       | |   override the variables or the simulation settings in the XML setup file
|                 | |       | | <-overrideFile=value> or <-overrideFile value>
|                 | |       | |   will override the variables or the simulation settings in the XML setup file with the values from the file
|                 | |       | | <-port=value> or <-port value>
|                 | |       | |   value specifies the port for simulation status (default disabled)
|                 | |       | | <-r=value> or <-r value>
|                 | |       | |   value specifies a new result file than the default Model_res.mat
|                 | |       | | <-reconcile>
|                 | |       | |   Run the Data Reconciliation numerical computation algorithm for constrained equations
|                 | |       | | <-reconcileBoundaryConditions>
|                 | |       | |   Run the Data Reconciliation numerical computation algorithm for boundary condition equations
|                 | |       | | <-reconcileState>
|                 | |       | |   Run the State Estimation numerical computation algorithm for constrained equations
|                 | |       | | <-gbm=value> or <-gbm value>
|                 | |       | |   Value specifies the chosen solver of solver gbode (single-rate, slow states integrator)
|                 | |       | | <-gbctrl=value> or <-gbctrl value>
|                 | |       | |   Step size control of solver gbode (single-rate, slow states integrator)
|                 | |       | | <-gberr=value> or <-gberr value>
|                 | |       | |   Error estimation method for solver gbode (single-rate, slow states integrator).
|                 | |       | | <-gbint=value> or <-gbint value>
|                 | |       | |   Interpolation method of solver gbode (single-rate, slow states integrator)
|                 | |       | | <-gbnls=value> or <-gbnls value>
|                 | |       | |   Non-linear solver method of solver gbode (single-rate, slow states integrator)
|                 | |       | | <-gbfm=value> or <-gbfm value>
|                 | |       | |   Value specifies the chosen solver of solver gbode (multi-rate, fast states integrator)
|                 | |       | | <-gbfctrl=value> or <-gbfctrl value>
|                 | |       | |   Step size control of solver gbode (multi-rate, fast states integrator)
|                 | |       | | <-gbferr=value> or <-gbferr value>
|                 | |       | |   Error estimation method for gbode solver (multi-rate, fast states integrator).
|                 | |       | | <-gbfint=value> or <-gbfint value>
|                 | |       | |   Interpolation method of solver gbode (multi-rate, fast states integrator)
|                 | |       | | <-gbfnls=value> or <-gbfnls value>
|                 | |       | |   Non-linear solver method of solver gbode (multi-rate, fast states integrator)
|                 | |       | | <-gbratio=value> or <-gbratio value>
|                 | |       | |   Define percentage of states for the fast states selection of solver gbode
|                 | |       | | <-rt=value> or <-rt value>
|                 | |       | |   value specifies the scaling factor for real-time synchronization (0 disables)
|                 | |       | | <-s=value> or <-s value>
|                 | |       | |   value specifies the integration method
|                 | |       | | <-single>
|                 | |       | |   output in single precision
|                 | |       | | <-steps>
|                 | |       | |   dumps the number of integration steps into the result file
|                 | |       | | <-steadyState>
|                 | |       | |   aborts if steady state is reached
|                 | |       | | <-steadyStateTol=value> or <-steadyStateTol value>
|                 | |       | |   [double (default 1e-3)] This relative tolerance is used to detect steady state.
|                 | |       | | <-sx=value> or <-sx value>
|                 | |       | |   value specifies a csv-file with inputs as covariance matrix Sx for DataReconciliation
|                 | |       | | <-keepHessian=value> or <-keepHessian value>
|                 | |       | |   value specifies the number of steps, which keep hessian matrix constant
|                 | |       | | <-w>
|                 | |       | |   shows all warnings even if a related log-stream is inactive
|                 | |       | | <-parmodNumThreads=value> or <-parmodNumThreads value>
|                 | |       | |   [int default: 0] value specifies the number of threads for simulation using parmodauto. If not specified (or is 0) it will use the systems max number of threads. Note that this option is ignored if the model is not compiled with --parmodauto
ERROR: failed process: Process(`/tmp/jl_Zs01Eq/BouncingBall -overrideFile=/tmp/jl_Zs01Eq/BouncingBall_override.txt '-noEventEmit -noRestart -override=e=0.3,g=9.71'`, ProcessExited(1)) [1]

Linearization

Missing docstring.

Missing docstring for linearize(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getLinearizationOptions(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for setLinearizationOptions(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getLinearInputs(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getLinearOutputs(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for getLinearStates(). Check Documenter's build log for details.

Examples

julia> getLinearizationOptions(mod)Dict{String, String} with 4 entries:
  "startTime" => "0.0"
  "stopTime"  => "1.0"
  "stepSize"  => "0.002"
  "tolerance" => "1e-6"
julia> getLinearizationOptions(mod, ["startTime","stopTime"])2-element Vector{String}:
 "0.0"
 "1.0"

julia> setLinearizationOptions(mod,["stopTime=2.0","tolerance=1e-06"])

julia> linearize(mod)LOG_STDOUT        | info    | Linearization will be performed at point of time: 2.000000
LOG_SUCCESS       | info    | The initialization finished successfully without homotopy method.
LOG_SUCCESS       | info    | The simulation finished successfully.
LOG_STDOUT        | info    | Linear model is created at /tmp/jl_Zs01Eq/linearized_model.jl
LOG_STDOUT        | info    | The output format can be changed with the command line option --linearizationDumpLanguage.
LOG_STDOUT        | info    | The options are: --linearizationDumpLanguage=modelica, matlab, julia, python.
ERROR: SystemError: opening file "/home/runner/work/OMJulia.jl/OMJulia.jl/docs/linearized_model.jl": No such file or directory

julia> getLinearInputs(mod)ERROR: UndefRefError: access to undefined reference
julia> getLinearOutputs(mod)ERROR: UndefRefError: access to undefined reference
julia> getLinearStates(mod)ERROR: UndefRefError: access to undefined reference

Sensitivity Analysis

OMJulia.sensitivity — Function

Method for computing numeric sensitivity of OpenModelica object

Arguments:

1st arg: Vp # Array of strings of Modelica Parameter names 2nd arg: Vv # Array of strings of Modelica Variable names 3rd arg: Ve # Array of float Excitations of parameters; defaults to scalar 1e-2

Returns:

1st return: VSname # Vector of Sensitivity names 2nd return: Sarray # Array of sensitivies: vector of elements per parameter, each element containing time series per variable

Examples

julia> (Sn, Sa) = sensitivity(mod, ["UA","EdR"], ["T","cA"], [1e-2,1e-4])ERROR: MethodError: no method matching parse(::Type{Float64}, ::Int64)

Closest candidates are:
  parse(::Type{T}, !Matched::AbstractString; kwargs...) where T<:Real
   @ Base parse.jl:384