Advanced API

A Julia style scripting API that handles low level API calls.

ModelicaSystem

OMJulia.ModelicaSystem — Function

ModelicaSystem(omc, filename, modelname, library=nothing;
               commandLineOptions=nothing, variableFilter=nothing)

Set command line options for OMCSession and build model modelname to prepare for a simulation.

Arguments

omc: OpenModelica compiler session, see OMCSession().

filename: Path to Modelica file.

modelname: Name of Modelica model to build, including namespace if the model is wrappen within a Modelica package.

library: List of dependent libraries or Modelica files. This argument can be passed as string (e.g. "Modelica") or tuple (e.g. ("Modelica", "4.0") or array (e.g. ["Modelica", "SystemDynamics"] or [("Modelica", "4.0"), "SystemDynamics"]).

Keyword Arguments

commandLineOptions: OpenModelica command line options, see OpenModelica Compiler Flags.

variableFilter: Regex to filter variables in result file.

Usage

using OMJulia
mod = OMJulia.OMCSession()
ModelicaSystem(mod, "BouncingBall.mo", "BouncingBall", ["Modelica", "SystemDynamics"], commandLineOptions="-d=newInst")

Providing dependent libaries:

using OMJulia
mod = OMJulia.OMCSession()
ModelicaSystem(mod, "BouncingBall.mo", "BouncingBall", ["Modelica", "SystemDynamics", "dcmotor.mo"])

See also OMCSession().

OMJulia.OMCSession — Type

OMCSession <: Any

OMC session struct.

OMCSession(omc=nothing)

Create new OpenModelica session.

Arguments

omc::Union{String, Nothing}: Path to OpenModelica compiler. Use omc from PATH if nothing is provided.

See also ModelicaSystem, OMJulia.quit.

OMJulia.quit — Function

quit(omc::OMCSession; timeout=4::Integer)

Quit OMCSession.

Arguments

- `omc::OMCSession`:      OMC session.

Keywords

- `timeout=4::Integer`:   Timeout in seconds.

See also OMJulia.OMCSession.

Example

Let us see the usage of ModelicaSystem with the help of Modelica model ModSeborgCSTRorg

model ModSeborgCSTRorg
  // Model of original Seborg CSTR in ode form
  // author: Bernt Lie, University of Southeast Norway,November 7, 2017

  // Parameters
  parameter Real V = 100 "Reactor volume, L";
  parameter Real rho = 1e3 "Liquid density, g/L";
  parameter Real a = 1 "Stoichiometric constant, -";
  parameter Real EdR = 8750 "Activation temperature, K";
  parameter Real k0 = exp(EdR/350) "Pre-exponential factor, 1/min";
  parameter Real cph = 0.239 "Specific heat capacity of mixture, J.g-1.K-1";
  parameter Real DrHt = -5e4 "Molar enthalpy of reaction, J/mol";
  parameter Real UA = 5e4 "Heat transfer parameter, J/(min.K)";

  // Initial state parameters
  parameter Real cA0 = 0.5 "Initial concentration of A, mol/L";
  parameter Real T0 = 350 "Initial temperature, K";
  // Declaring variables
  // -- states
  Real cA(start = cA0, fixed = true) "Initializing concentration of A in reactor, mol/L";
  Real T(start = T0, fixed = true) "Initializing temperature in reactor, K";
  // -- auxiliary variables
  Real r "Rate of reaction, mol/(L.s)";
  Real k "Reaction 'constant', ...";
  Real Qd "Heat flow rate, J/min";
  // -- input variables
  input Real Vdi "Volumetric flow rate through reactor, L/min";
  input Real cAi "Influent molar concentration of A, mol/L";
  input Real Ti "Influent temperature, K";
  input Real Tc "Cooling temperature', K";
  // -- output variables
  output Real y_T "Reactor temperature, K";
  // Equations constituting the model
equation
  // Differential equations
  der(cA) = Vdi*(cAi-cA)/V- a*r;
  der(T) = Vdi*(Ti-T)/V + (-DrHt)*r/(rho*cph) + Qd/(rho*V*cph);
  // Algebraic equations
  r = k*cA^a;
  k = k0*exp(-EdR/T);
  Qd = UA*(Tc-T);
  // Outputs
  y_T = T;
end ModSeborgCSTRorg

julia> using OMJulia
julia> mod = OMJulia.OMCSession()[ Info: Path to zmq file="/tmp/openmodelica.runner.port.julia.fZRG4JxRwS"
OMJulia.OMCSession(false, false, Dict{Any, Any}(), Dict{Any, Any}(), Dict{Any, Any}(), "", "/home/runner/work/OMJulia.jl/OMJulia.jl", "", "", "", "", "", nothing, Any[], Dict{Any, Any}(), Dict{Any, Any}(), Dict{Any, Any}(), Dict{Any, Any}(), OMJulia.ZMQSession(ZMQ.Context(Ptr{Nothing} @0x00000000028c53e0, WeakRef[WeakRef(ZMQ.Socket(Ptr{Nothing} @0x00000000029d9100, FileWatching._FDWatcher(Ptr{Nothing} @0x000000000250fe90, 30, (1, 0), Base.GenericCondition{Base.Threads.SpinLock}(Base.IntrusiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0)), 0, (false, false))))]), ZMQ.Socket(Ptr{Nothing} @0x00000000029d9100, FileWatching._FDWatcher(Ptr{Nothing} @0x000000000250fe90, 30, (1, 0), Base.GenericCondition{Base.Threads.SpinLock}(Base.IntrusiveLinkedList{Task}(nothing, nothing), Base.Threads.SpinLock(0)), 0, (false, false))), Process(`omc --interactive=zmq -z=julia.fZRG4JxRwS`, ProcessRunning)), OMJulia.Linearization("", "", Dict{AbstractString, AbstractString}("startTime" => "0.0", "stopTime" => "1.0", "stepSize" => "0.002", "tolerance" => "1e-6"), false, missing, missing, missing))
julia> ModelicaSystem(mod,
                      joinpath("docs", "testmodels", "ModSeborgCSTRorg.mo"),
                      "ModSeborgCSTRorg")ERROR: Error: Error building simulator. Build log: clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg.o ModSeborgCSTRorg.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_functions.o ModSeborgCSTRorg_functions.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_records.o ModSeborgCSTRorg_records.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_01exo.o ModSeborgCSTRorg_01exo.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_02nls.o ModSeborgCSTRorg_02nls.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_03lsy.o ModSeborgCSTRorg_03lsy.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_04set.o ModSeborgCSTRorg_04set.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_05evt.o ModSeborgCSTRorg_05evt.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_06inz.o ModSeborgCSTRorg_06inz.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_07dly.o ModSeborgCSTRorg_07dly.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_08bnd.o ModSeborgCSTRorg_08bnd.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_09alg.o ModSeborgCSTRorg_09alg.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_10asr.o ModSeborgCSTRorg_10asr.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_11mix.o ModSeborgCSTRorg_11mix.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_12jac.o ModSeborgCSTRorg_12jac.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_13opt.o ModSeborgCSTRorg_13opt.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_14lnz.o ModSeborgCSTRorg_14lnz.c
ModSeborgCSTRorg_14lnz.c:31:1: warning: non-void function does not return a value [-Wreturn-type]
}
^
1 warning generated.
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_15syn.o ModSeborgCSTRorg_15syn.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_16dae.o ModSeborgCSTRorg_16dae.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_17inl.o ModSeborgCSTRorg_17inl.c
clang  -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0  -c -o ModSeborgCSTRorg_18spd.o ModSeborgCSTRorg_18spd.c
clang -I. -o ModSeborgCSTRorg ModSeborgCSTRorg.o ModSeborgCSTRorg_functions.o ModSeborgCSTRorg_records.o ModSeborgCSTRorg_01exo.o ModSeborgCSTRorg_02nls.o ModSeborgCSTRorg_03lsy.o ModSeborgCSTRorg_04set.o ModSeborgCSTRorg_05evt.o ModSeborgCSTRorg_06inz.o ModSeborgCSTRorg_07dly.o ModSeborgCSTRorg_08bnd.o ModSeborgCSTRorg_09alg.o ModSeborgCSTRorg_10asr.o ModSeborgCSTRorg_11mix.o ModSeborgCSTRorg_12jac.o ModSeborgCSTRorg_13opt.o ModSeborgCSTRorg_14lnz.o ModSeborgCSTRorg_15syn.o ModSeborgCSTRorg_16dae.o ModSeborgCSTRorg_17inl.o ModSeborgCSTRorg_18spd.o -L"/home/runner/work/OMJulia.jl/OMJulia.jl/docs/testmodels"    -Os -fPIC -falign-functions -mfpmath=sse -fno-dollars-in-identifiers -Wno-parentheses-equality     -I"/usr/bin/../include/omc/c" -I"/usr/bin/../include/omc" -I. -DOPENMODELICA_XML_FROM_FILE_AT_RUNTIME -DOMC_MODEL_PREFIX=ModSeborgCSTRorg -DOMC_NUM_MIXED_SYSTEMS=0 -DOMC_NUM_LINEAR_SYSTEMS=0 -DOMC_NUM_NONLINEAR_SYSTEMS=0 -DOMC_NDELAY_EXPRESSIONS=0 -DOMC_NVAR_STRING=0 -L"/usr/bin/../lib/x86_64-linux-gnu/omc" -L"/usr/bin/../lib" -Wl,-rpath,"/usr/bin/../lib/x86_64-linux-gnu/omc"  -Wl,-rpath,"/usr/bin/../lib"   -Wl,--no-as-needed -Wl,--disable-new-dtags -lSimulationRuntimeC -llapack -lblas  -lm -lomcgc -lryu -lpthread -rdynamic -Wl,--no-undefined
/usr/bin/ld: cannot find -lryu: No such file or directory
clang: error: linker command failed with exit code 1 (use -v to see invocation)
make: *** [ModSeborgCSTRorg.makefile:37: omc_main_target] Error 1

WorkDirectory

For each OMJulia.OMCSession session a temporary work directory is created and the results are published in that working directory. In order to get the work directory use getWorkDirectory.

OMJulia.getWorkDirectory — Function

getWorkDirectory(omc)

Return working directory of OMJulia.OMCsession omc.

julia> getWorkDirectory(mod)"/tmp/jl_1j4R3i"

Build Model

OMJulia.buildModel — Function

buildModel(omc; variableFilter=nothing)

Build modelica model.

Arguments

omc::OMCSession: OpenModelica compiler session.

Keyword Arguments

variableFilter: Regex to filter variables in result file.

In case the Modelica model needs to be updated or additional simulation flags needs to be set using sendExpression The buildModel API can be used after ModelicaSystem.

buildModel(omc)
buildModel(omc, variableFilter="a|T")

Get Methods

OMJulia.getQuantities — Function

getQuantities(omc, name=nothing)

Return list of all variables parsed from xml file.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of variables to read from xml file. If nothing is provided read all variables.

See also showQuantities.

OMJulia.showQuantities — Function

showQuantities(omc, name=nothing)

Return DataFrame of all variables parsed from xml file.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of variables to read from xml file. If nothing is provided read all variables.

See also getQuantities.

OMJulia.getContinuous — Function

getContinuous(omc, name=nothing)

Return continuous variables parsed from xml file.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of continuous variables to read from xml file. If nothing is provided read all continuous variables.

OMJulia.getInputs — Function

getInputs(omc, name=nothing)

Return input variables parsed from xml file. If input variables have no start value the returned value is "None".

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of input variables to read from xml file. If nothing is provided read all input variables.

OMJulia.getOutputs — Function

getInputs(omc, name=nothing)

Return output variables parsed from xml file. If output variables have no start value the returned value is "None".

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of output variables to read from xml file. If nothing is provided read all output variables.

OMJulia.getParameters — Function

getParameters(omc, name=nothing)

Return parameter variables parsed from xml file.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of parameters to read from xml file. If nothing is provided read all parameters.

OMJulia.getSimulationOptions — Function

getSimulationOptions(omc, name=nothing)

Return SimulationOption variables parsed from xml file.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of parameters to read from xml file. If nothing is provided read all parameters.

OMJulia.getSolutions — Function

getSolutions(omc::OMCSession, name=nothing; resultfile=nothing)

Read result file and return simulation results

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of variables to read from result file. If nothing is provided read all variables.

Keyword Arguments

resultfile::Union{AbstractString, Nothing}: Path to result file. If nothing is provided use saved result file.

Examples

julia> getQuantities(mod)Any[]
julia> getQuantities(mod, "T")Any[]
julia> getQuantities(mod, ["T","cA"])Any[]
julia> showQuantities(mod)ERROR: BoundsError: attempt to access 0-element Vector{Any} at index [1]

julia> getContinuous(mod)Dict{Any, Any}()
julia> getContinuous(mod, ["Qd","Tc"])2-element Vector{Int64}:
 0
 0

julia> getInputs(mod)Dict{Any, Any}()
julia> getOutputs(mod)Dict{Any, Any}()

julia> getParameters(mod)Dict{Any, Any}()
julia> getParameters(mod, ["a","V"])2-element Vector{Int64}:
 0
 0

julia> getSimulationOptions(mod)Dict{Any, Any}()
julia> getSimulationOptions(mod, ["stepSize","tolerance"])2-element Vector{Int64}:
 0
 0

Reading Simulation Results

To read the simulation results, we need to simulate the model first and use the getSolution() API to read the results

julia> simulate(mod)

The getSolution method can be used in two different ways.

using default result filename

use the result filenames provided by user

This provides a way to compare simulation results and perform regression testing

julia> getSolutions(mod)ERROR: Result file /tmp/jl_1j4R3i/ModSeborgCSTRorg_res.mat does not exist !
julia> getSolutions(mod, ["time","a"])ERROR: Result file /tmp/jl_1j4R3i/ModSeborgCSTRorg_res.mat does not exist !

Examples of using resultFile provided by user location

getSolutions(mod, resultfile="C:/BouncingBal/tmpbouncingBall.mat") //returns list of simulation variables for which results are available , the resulfile location is provided by user
getSolutions(mod, ["time","h"], resultfile="C:/BouncingBal/tmpbouncingBall.mat") // return list of array

Set Methods

OMJulia.setInputs — Function

setInputs(omc, name)

Set new values for input variables.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}}: String "Name=value" or vector of strings ["Name1=value1","Name2=value2","Name3=value3"])

OMJulia.setParameters — Function

setParameters(omc, name; verbose=true)

Set parameter values for parameter variables defined by users

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}}: String "Name=value" or vector of strings ["Name1=value1","Name2=value2","Name3=value3"])

Keyword Arguments

verbose::Bool: Display additional info if setParameters failed.

OMJulia.setSimulationOptions — Function

setSimulationOptions(omc, name)

Set simulation option values like stopTime or stepSize.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}}: String "Name=value" or vector of strings ["Name1=value1","Name2=value2","Name3=value3"])

Examples

julia> setInputs(mod, "cAi=100")ERROR: cAi is not an input variable
julia> setInputs(mod, ["cAi=100","Ti=200","Vdi=300","Tc=250"])ERROR: cAi is not an input variable

julia> setParameters(mod, "a=3")[ Info: setParameters() failed: " a" is not a parameter
julia> setParameters(mod, ["a=4","V=200"])[ Info: setParameters() failed: " a" is not a parameter
[ Info: setParameters() failed: " V" is not a parameter

julia> setSimulationOptions(mod, ["stopTime=2.0", "tolerance=1e-08"])ERROR: "stopTime" is not a simulation option

Advanced Simulation

OMJulia.simulate — Function

simulate(omc; resultfile=nothing, simflags="", verbose=false)

Simulate modelica model.

Arguments

omc::OMCSession: OpenModelica compiler session, see OMCSession().

Keyword Arguments

resultFile::Union{String, Nothing}: Result file to write simulation results into.

simflags::String: Simulation flags, see Simulation Runtime Flags.

verbose::Bool: [debug] Log cmd call to log.txt and error.txt.

Examples

simulate(omc)

Specify result file:

simulate(omc, resultfile="tmpresult.mat")

Set simulation runtime flags:

simulate(omc, simflags="-noEmitEvent -override=e=0.3,g=9.3")

An example of how to do advanced simulation to set parameter values using set methods and finally simulate the "ModSeborgCSTRorg.mo" model is given below .

julia> getParameters(mod)Dict{Any, Any}()
julia> setParameters(mod, "a=3.0")[ Info: setParameters() failed: " a" is not a parameter

To check whether new values are updated to model , we can again query the getParameters().

julia> getParameters(mod)Dict{Any, Any}()

Similary we can also use setInputs() to set a value for the inputs during various time interval can also be done using the following.

julia> setInputs(mod, "cAi=100")ERROR: cAi is not an input variable

And finally we simulate the model

julia> simulate(mod)

Linearization

OMJulia.linearize — Function

function which returns the linearize model of modelica model, The function returns four matrices A, B, C, D

linearize(omc; lintime = nothing, simflags= nothing, verbose=true)

Arguments

omc::OMCSession: OpenModelica compiler session.

Keyword Arguments

lintime : Value specifies a time where the linearization of the model should be performed

simflags: Simulation flags, see Simulation Runtime Flags.

Examples of using linearize() API

linearize(omc)

Specify result file:

linearize(omc, lintime="0.5")

Set simulation runtime flags:

linearize(omc, simflags="-noEmitEvent")

OMJulia.getLinearizationOptions — Function

getLinearizationOptions(omc, name=nothing)

Return linearization options.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}, Nothing}: Names of linearization options. If nothing is provided return all linearization options.

OMJulia.setLinearizationOptions — Function

setLinearizationOptions(omc, name)

Set linearization options.

Arguments

omc::OMCSession: OpenModelica compiler session.

name::Union{<:AbstractString, Array{<:AbstractString,1}}: String "Name=value" or vector of strings ["Name1=value1","Name2=value2","Name3=value3"])

OMJulia.getLinearInputs — Function

getLinearInputs(omc)

Return linear input variables after the model is linearized

Arguments

omc::OMCSession: OpenModelica compiler session.

OMJulia.getLinearOutputs — Function

getLinearOutputs(omc)

Return linear output variables after the model is linearized

Arguments

omc::OMCSession: OpenModelica compiler session.

OMJulia.getLinearStates — Function

getLinearStates(omc)

Return linear state variables after the model is linearized

Arguments

omc::OMCSession: OpenModelica compiler session.

Examples

julia> getLinearizationOptions(mod)Dict{AbstractString, AbstractString} with 4 entries:
  "startTime" => "0.0"
  "stopTime"  => "1.0"
  "stepSize"  => "0.002"
  "tolerance" => "1e-6"
julia> getLinearizationOptions(mod, ["startTime","stopTime"])2-element Vector{String}:
 "0.0"
 "1.0"

julia> setLinearizationOptions(mod,["stopTime=2.0","tolerance=1e-06"])

julia> res = linearize(mod)ERROR: Linearization cannot be performed as the model is not build, use ModelicaSystem() to build the model first

julia> getLinearInputs(mod)ERROR: Model is not linearized
julia> getLinearOutputs(mod)Model is not Linearized
julia> getLinearStates(mod)Model is not Linearized

Sensitivity Analysis

OMJulia.sensitivity — Function

sensitivity(omc::OMCSession, Vp, Vv, Ve=[1e-2])

Method for computing numeric sensitivity of OpenModelica object.

Arguments

omc::OMCSession: OpenModelica compiler session.

Vp::Array{<:AbstractString, 1}: Modelica Parameter names.

Vv::Array{<:AbstractString, 1}: Modelica Variable names.

Ve::Array{Float64, 1}: Excitations of parameters; defaults to scalar 1e-2

Return

VSname::Vector{Vector{String}}: Vector of sensitivity names

VSarray::Vector{Vector{Vector{Float64}}}: Vector of sensitivies: vector of elements per parameter

Each element containing time series per variable

Examples

julia> (Sn, Sa) = sensitivity(mod, ["UA","EdR"], ["T","cA"], [1e-2,1e-4])ERROR: MethodError: no method matching parse(::Type{Float64}, ::Int64)

Closest candidates are:
  parse(::Type{T}, !Matched::AbstractString; kwargs...) where T<:Real
   @ Base parse.jl:384